Move methods used exclusively in CfWA11YTest to CfWA11YTest
